+++ src/config/config.go
@@ -50,6 +50,9 @@ const (
 	DefaultSignalAddr           = "127.0.0.1:2443"
 	DefaultSignalRealm          = "main"
 	DefaultSignalSkipVerify     = false
+	DefaultICEAddress           = "stun:stun.l.google.com:19302"
+	DefaultICEUsername          = ""
+	DefaultICEPassword          = ""
 )
 
 // Config contains all the configuration properties of a Babble node.
@@ -167,10 +170,20 @@ type Config struct {
 	// attacks. This should be used only for testing.
 	SignalSkipVerify bool `mapstructure:"signal-skip-verify"`
 
-	// ICEServers defines a slice describing servers available to be used by
-	// ICE, such as STUN and TURN servers.
+	// ICE address is the URL of a server providing services for ICE, such as
+	// STUN and TURN. The server should support password-based authentication,
+	// as Babble will try to connect with the username and password provided in
+	// ICEUsername and ICEPassword below.
 	// https://developer.mozilla.org/en-US/docs/Web/API/RTCIceServer/urls
-	ICEServers []webrtc.ICEServer
+	ICEAddress string `mapstructure:"ice-addr"`
+
+	// ICEUsername is the username that will be used to authenticate with the
+	// ICE server defined in ICEAddress.
+	ICEUsername string `mapstructure:"ice-username"`
+
+	// ICEPassword is the password that will be used to authenticate with the
+	// ICE server defined in ICEAddress.
+	ICEPassword string `mapstructure:"ice-password"`
 
 	// Proxy is the application proxy that enables Babble to communicate with
 	// the application.
@@ -207,7 +220,9 @@ func NewDefaultConfig() *Config {
 		SignalAddr:           DefaultSignalAddr,
 		SignalRealm:          DefaultSignalRealm,
 		SignalSkipVerify:     DefaultSignalSkipVerify,
-		ICEServers:           DefaultICEServers(),
+		ICEAddress:           DefaultICEAddress,
+		ICEUsername:          DefaultICEUsername,
+		ICEPassword:          DefaultICEPassword,
 	}
 
 	return config
@@ -241,6 +256,21 @@ func (c *Config) Keyfile() string {
 // certificate.
 func (c *Config) CertFile() string {
 	return filepath.Join(c.DataDir, DefaultCertFile)
+}
+
+// ICEServers returns a list of ICE servers used by the WebRTCStreamLayer to
+// connect to peers. The list contains a single item which is based on the
+// configuration passed through the config object. This configuration is limited
+// to a single server, with password-based authentication.
+func (c *Config) ICEServers() []webrtc.ICEServer {
+	return []webrtc.ICEServer{
+		{
+			URLs:           []string{c.ICEAddress},
+			Username:       c.ICEUsername,
+			Credential:     c.ICEPassword,
+			CredentialType: webrtc.ICECredentialTypePassword,
+		},
+	}
 }
 
 // Logger returns a formatted logrus Entry, with prefix set to "babble".
@@ -307,13 +337,12 @@ func LogLevel(l string) logrus.Level {
 	}
 }
 
-// DefaultICEServers returns a list containing a single ICEServer which
-// points to a public STUN server provided by Google. This default configuration
-// does not include a TURN server, so not all p2p connections will be possible.
+// DefaultICEServers returns the default ICE configuration with one URL pointing
+// to a public Google STUN server.
 func DefaultICEServers() []webrtc.ICEServer {
 	return []webrtc.ICEServer{
 		{
-			URLs: []string{"stun:stun.l.google.com:19302"},
+			URLs: []string{DefaultICEAddress},
 		},
 	}
 }